refactor(suite_core): move Rate Limit doctype and rate limiter util to Suite Core
The Rate Limit DocType is common across modules, so move it from Mail to
Suite Core and relocate the dynamic rate limiter helper to the shared
suite.utils package. Add a pre_model_sync patch to retag the DocType's
module on upgraded sites.
