+++ suite/mail/api/auth.py
@@ -2,7 +2,7 @@ import frappe
 from frappe import _
 
 from suite.mail.doctype.user_account.user_account import get_user_jmap_accounts
-from suite.mail.utils.rate_limiter import dynamic_rate_limit
+from suite.utils.rate_limiter import dynamic_rate_limit
 from suite.mail.utils.user import get_account_emails, is_jmap_configured
