+++ suite/patches.txt
@@ -19,6 +19,8 @@ suite.mail.patches.delete_client_module
 suite.mail.patches.capture_user_outgoing_settings
 suite.mail.patches.rebuild_rate_limit_unique_constraint
 suite.mail.patches.rename_account_settings_to_jmap_account
+# --- suite_core ---
+suite.suite_core.patches.move_rate_limit_to_suite_core
 [post_model_sync]
 # --- drive ---
 suite.drive.patches.folder_size #3
