+++ suite/suite_core/doctype/rate_limit/rate_limit.js
