+++ suite/utils/rate_limiter.py
@@ -4,7 +4,7 @@ import frappe
 from frappe import _
 from frappe.rate_limiter import rate_limit
 
-from suite.mail.doctype.rate_limit.rate_limit import get_rate_limits
+from suite.suite_core.doctype.rate_limit.rate_limit import get_rate_limits
 
 
 def dynamic_rate_limit() -> callable:
